wr-95: * move to shared
