+++ backend/src/libs/packages/file/libs/enums/enums.ts
@@ -1,1 +1,1 @@
-export { FileType } from './file-type.enum.js';
+export { FileType } from 'shared/build/index.js';
