+++ shared/src/packages/files/files.ts
@@ -1,4 +1,8 @@
-export { FilesApiPath } from './libs/enums/enums.js';
+export {
+  MAX_FILE_SIZE_MB,
+  SUPPORTED_FILE_TYPES,
+} from './libs/constants/constants.js';
+export { FilesApiPath, FileType } from './libs/enums/enums.js';
 export {
   type FileUploadRequestDto,
   type FileUploadResponseDto,
