+++ shared/src/packages/files/libs/enums/enums.ts
@@ -1,1 +1,2 @@
+export { FileType } from './file-type.enum.js';
 export { FilesApiPath } from './files-api-path.enum.js';
